move checkpoint api to trainer.py
